credit_system/requirements.txt
asgiref==3.9.2
Django==5.2.6
django-environ==0.12.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
